+++ solana/types/tokenProgramAction.go
@@ -4,7 +4,7 @@ type TokenProgramTransferAction struct {
 	BaseAction
 	From   string `json:"from"`
 	To     string `json:"to"`
-	Amount uint32 `json:"amount"`
+	Amount uint64 `json:"amount"`
 }
 
 type TokenProgramTransferCheckedAction struct {
